# cmd/docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GoFundIt is a collaborative platform enabling families and friends to plan and fund group activities, such as vacations and events. Features end-to-end encryption, secure payment handling, and flexible contribution options.",
        "title": "GoFundIt API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://gofundit.com/support",
            "email": "oyeniyibright@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/payout/bank-list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves list of available banks for payout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payout"
                ],
                "summary": "Get Bank List",
                "responses": {
                    "200": {
                        "description": "Bank list retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/paystack.Bank"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/handlers.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/payout/manual/{campaignID}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Initializes a manual payout for a campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payout"
                ],
                "summary": "Initialize Manual Payout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Manual payout details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PayoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Manual payout initialized successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Payout"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid payout details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handlers.ValidationError"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Campaign not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.response"
                        }
                    }
                }
            }
        },
        "/payout/verify/bank-account": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Verifies a bank account for payout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payout"
                ],
                "summary": "Verify Bank Account",
                "parameters": [
                    {
                        "description": "Account verification details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VerifyAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account verified successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/paystack.ResolveAccountResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid account details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handlers.ValidationError"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/payout/{campaignID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves payout information for a campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payout"
                ],
                "summary": "Get Campaign Payout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payout information retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Payout"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Campaign not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Initializes a payout for a campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payout"
                ],
                "summary": "Initialize Payout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payout details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PayoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payout initialized successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Payout"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid payout details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handlers.ValidationError"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Campaign not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.PayoutRequest": {
            "type": "object",
            "required": [
                "accountName",
                "accountNumber",
                "bankCode",
                "bankName"
            ],
            "properties": {
                "accountName": {
                    "type": "string",
                    "minLength": 3
                },
                "accountNumber": {
                    "type": "string"
                },
                "bankCode": {
                    "type": "string"
                },
                "bankName": {
                    "type": "string",
                    "minLength": 3
                }
            }
        },
...
        "dto.VerifyAccountRequest": {
            "type": "object",
            "required": [
                "accountNumber",
                "bankCode"
            ],
            "properties": {
                "accountNumber": {
                    "type": "string"
                },
                "bankCode": {
                    "type": "string"
                }
            }
        },
...
        "handlers.BadRequestResponse": {
            "description": "Bad request error response structure",
            "type": "object",
            "properties": {
                "errors": {
                    "description": "Validation errors (optional)\n@example [{\"field\":\"field\",\"error\":\"must be a valid data\"}]"
                },
                "message": {
                    "description": "Error message\n@example Invalid input provided",
                    "type": "string"
                },
                "status": {
                    "description": "Status will always be \"Bad Request\"\n@example Bad Request",
                    "type": "string",
                    "enum": [
                        "Bad Request"
                    ],
                    "example": "Bad Request"
                }
            }
        },
        "handlers.SuccessResponse": {
            "description": "Successful API response structure",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data (optional)\n@example null"
                },
                "message": {
                    "description": "Response message describing the success\n@example Operation completed successfully",
                    "type": "string"
                },
                "status": {
                    "description": "Status will always be \"OK\"\n@example OK",
                    "type": "string",
                    "enum": [
                        "OK"
                    ],
                    "example": "OK"
                }
            }
        },
        "handlers.UnauthorizedResponse": {
            "description": "Unauthorized error response structure",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Will always be \"Unauthorized\"\n@example Unauthorized",
                    "type": "string",
                    "enum": [
                        "Unauthorized"
                    ],
                    "example": "Unauthorized"
                },
                "status": {
                    "description": "Status will always be false for unauthorized\n@example false",
                    "type": "string",
                    "enum": [
                        "Unauthorized"
                    ],
                    "example": "Unauthorized"
                }
            }
        },
        "handlers.ValidationError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "must be a valid field"
                },
                "field": {
                    "type": "string",
                    "example": "field"
                }
            }
        },
        "handlers.response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data (optional)"
                },
                "errors": {
                    "description": "Error details (optional)"
                },
                "message": {
                    "description": "Response message\n@example Operation completed successfully",
                    "type": "string"
                },
                "status": {
                    "description": "Response status\n@example OK",
                    "type": "string"
                }
            }
        },
...
        "models.CryptoAccount": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cryptoToken": {
                    "$ref": "#/definitions/models.CryptoToken"
                }
            }
        },
        "models.CryptoToken": {
            "type": "string",
            "enum": [
                "USDT",
                "USDC",
                "BUSD",
                "DAI"
            ],
            "x-enum-varnames": [
                "USDT",
                "USDC",
                "BUSD",
                "DAI"
            ]
        },
        "models.FiatAccount": {
            "type": "object",
            "properties": {
                "accountName": {
                    "type": "string"
                },
                "accountNumber": {
                    "type": "string"
                },
                "bankCode": {
                    "type": "string"
                },
                "bankName": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                }
            }
        },
...
        "models.PaymentMethod": {
            "type": "string",
            "enum": [
                "fiat",
                "crypto",
                "manual"
            ],
            "x-enum-varnames": [
                "PaymentMethodFiat",
                "PaymentMethodCrypto",
                "PaymentMethodManual"
            ]
        },
...
        "models.Payout": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "campaignId": {
                    "type": "string"
                },
                "completedAt": {
                    "type": "string"
                },
                "cryptoAccount": {
                    "$ref": "#/definitions/models.CryptoAccount"
                },
                "failureReason": {
                    "type": "string"
                },
                "fiatAccount": {
                    "$ref": "#/definitions/models.FiatAccount"
                },
                "payoutMethod": {
                    "$ref": "#/definitions/models.PaymentMethod"
                },
                "processedAt": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.PayoutStatus"
                }
            }
        },
        "models.PayoutStatus": {
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "completed",
                "failed"
            ],
            "x-enum-varnames": [
                "PayoutStatusPending",
                "PayoutStatusProcessing",
                "PayoutStatusCompleted",
                "PayoutStatusFailed"
            ]
        },
...
        "paystack.Bank": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "paystack.ResolveAccountResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "account_name": {
                            "type": "string"
                        },
                        "account_number": {
                            "type": "string"
                        }
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for accessing the API endpoints",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT token for authenticated requests. Use the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CampaignKeyAuth": {
            "description": "Campaign Key required for campaign-specific operations",
            "type": "apiKey",
            "name": "Campaign-Key",
            "in": "header"
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ],
...
}